refactor(radio): extract recording_commands + add deferred #5 logging
GATE-11 decomposition of recording.py (707 -> 563, back under the 600 default):
the pure ffmpeg/ffprobe command + filename builders (build_record_command,
build_probe_codec_command, parse_probe_codec, raw_capture_extension,
build_filename + the RECORD_FORMATS/_CODECS/_RAW_* constants) move to a new
quill/core/radio/recording_commands.py. recording.py re-exports them, so
`from ...recording import build_record_command` keeps working; its budget entry
is removed (now under cap).

With the headroom, the deferred quill-radio #5 logging lands:
- recording finalization: how a recording ended and where it finalized.
- playback-engine selection: preference + mpv-present -> chosen backend
  (player_controller, +7, rebaselined 605 -> 612).

No behavior change; 274 radio tests pass, ruff/mypy clean, GATE-11 OK.
