+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Improved
 
+- **More radio debug logging: playback-engine selection and recording finalization (quill-radio #5).** With radio debug mode on, `quill.log` now also records which playback backend was chosen for a station and why (preference + whether mpv is present), and how each recording ended and where its finished file landed. Internal cleanup rode along: the pure ffmpeg/ffprobe command and filename builders moved out of `recording.py` into a new `recording_commands` module (no behavior change; the recorder module dropped back under its size budget).
+
 - **Radio stream discovery now leaves a debug trail (quill-radio #5).** With radio debug mode on, Find Streams from a Website logs each scanned page and the candidate links it kept (with the reason for each, including the iHeart/TuneIn portal-page follow), and the Triton/StreamTheWorld resolver logs which callsign resolved to how many playable streams -- so a "why didn't this station's stream show up?" question has an answer in `quill.log`. Every logged URL is redacted first. (More per-module coverage -- engine selection, reconnects, recording finalization -- follows with the recording-module decomposition.)
 
 - **Radio recordings can use a temporary folder while in progress (quill-radio #5).** Recording Settings gains an optional **Temporary folder (while recording)**: when set, a recording is written there and moved to your destination folder the moment it finishes, so a half-written file never appears in your recordings folder and a fast scratch/SSD volume can absorb the write churn. The move is atomic on the same volume and falls back to copy-then-delete across volumes; if the move ever fails the finished file is left safely in the temp folder rather than lost, and an unusable temp folder simply records straight to the destination. Blank keeps today's behavior (record straight to the destination).
