+++ quill/tools/module_size_budgets.json
@@ -350,9 +350,6 @@
   "_rebaseline_2026_06_27_prompt_store_consolidation": "quill/__main__.py 638->640 and quill/core/prompt_library.py 364->393 (+29) for the one-prompt-store-truth migration (AI Library Phase 2/3, docs/planning/ai-menu-redesign-plan.md). __main__.py: the two-line guarded startup call (import + consolidate_prompts_quietly()) alongside the existing consolidate_provider_keys_quietly() hook; the migration logic itself lives in the new wx-free quill/core/prompt_migration.py so it adds no growth to either store. prompt_library.py: the new upsert_external() method, which inserts/replaces a user prompt by explicit id (refusing built-ins) so import and the store migration are idempotent and reversible. Same accepted startup-hook / additive-method rebaseline pattern. Budgets re-baselined to current size.",
   "_rebaseline_2026_06_27_main_resync_merge": "Merge of origin/main (0.8.0 line: Python 3.13 runtime, upgrade hardening, legacy-import + startup-maintenance, winrt OCR, bundled Kokoro) into 2.0-dev. Seven modules touched independently on both branches now hold the union of both sides' additions, so the max-of-both budget no longer fits: __main__.py 640->651 (both startup hooks: apply_pending_legacy_import + run_pending_startup_maintenance from main, consolidate_provider_keys/prompts from 2.0-dev), keymap.py 682->683, settings.py 1226->1231, ai_hub_dialog.py 924->927, assistant_panel.py 943->947, main_frame.py 26172->26218, main_frame_statusbar.py 956->967. These are merge combinations of pre-approved per-branch growth, not new feature work; CQ-1 decomposition remains the path down for the large UI modules. Budgets re-baselined to current size.",
   "budgets": {
-    "quill/core/radio/recording.py": 707,
-    "_rebaseline_2026_07_17_radio_recording_stderr_capture": "core/radio/recording.py enters the tracked list at 636 (was under the 600 default cap): quill-radio #4/#5, capturing ffmpeg's live recording stderr. The +36 over the cap is the _drain_stderr daemon-thread method (reads ffmpeg's stderr line by line, redacts, logs error-shaped lines at WARNING and the rest at DEBUG), the _FFMPEG_LOGLEVELS/_STDERR_ERROR_RE module constants, the loglevel param on build_record_command, and the stderr=PIPE + drain-thread spawn in start(). Cohesive with the recorder's existing start/monitor/reconnect lifecycle already in this file; the logging switch it responds to is the under-cap core/radio/radio_logging.",
-    "_rebaseline_2026_07_17_radio_recording_temp_dir": "recording.py 636 -> 707 (+71): quill-radio #5, record-to-temp then move-to-final. RecordingSettings.temp_dir field (+ to_dict/from_dict), the temp-root resolution in start() (mkdir with a fall-back-to-destination-on-error), the _final_destination tracking through start()/_monitor(), and the new module-level _finalize_move() helper (os.replace on the same volume, copy+delete across volumes, leave-in-temp on failure so a finished recording is never lost). DECOMPOSITION DUE: this module is now the largest core/radio file; the pure command/filename builders (build_record_command, build_probe_codec_command, parse_probe_codec, raw_capture_extension, build_filename + the RECORD_FORMATS/_CODECS/_RAW_* constants) should be extracted into a recording_commands.py in a dedicated follow-up so the ratchet can resume decreasing.",
     "quill/apps/radio.py": 1081,
     "_rebaseline_2026_07_17_radio_missed_recording_report": "apps/radio.py 1051 -> 1081 (+30): quill-radio #4, the startup missed-recording report. _report_missed_recordings (scheduled via wx.CallAfter at launch) reads RadioHistory.last_seen, asks the wx-free recording_schedule.missed_occurrences/describe_missed which enabled schedules fired while the app was closed, announces a one-line summary, and re-stamps last_seen; the exit path also stamps last_seen=now so the next launch's window is the closed gap. The occurrence math + the spoken-summary formatter live in core/radio/recording_schedule.py (under cap); last_seen lives in core/radio/history.py.",
     "_rebaseline_2026_07_17_radio_debug_mode_pref": "apps/radio.py 1038 -> 1051 (+13): quill-radio #5, the Preferences (Ctrl+,) 'Verbose logging (debug mode)' checkbox -- one PreferenceCheckbox spec, its addition to the checkbox-values unpack, and the immediate set_radio_debug(history.debug_mode) apply-on-change block so it takes effect this session. The debug_mode field lives in core/radio/history.py and the logging switch itself is the under-cap core/radio/radio_logging.set_radio_debug.",
@@ -365,7 +362,8 @@
     "quill/ui/main_frame_radio.py": 1175,
     "_rebaseline_2026_07_17_radio_debug_mode_apply": "main_frame_radio.py 1171 -> 1175 (+4): quill-radio #5, _init_radio now applies the persisted verbose-logging preference at startup -- a two-line import + set_radio_debug(self._radio_history.debug_mode) call with a comment -- so debug mode is in effect for both embedded QUILL radio and the standalone (which share this mixin). The switch itself is the under-cap core/radio/radio_logging.set_radio_debug.",
     "_rebaseline_2026_07_17_radio_now_playing_free_fallback": "main_frame_radio.py 1144 -> 1171 (+27): #1111 free 'What's Playing' fallback. When ICY and the player's in-band title both give nothing, _radio_fetch_track_title now delegates to a new _radio_fetch_server_now_playing that reads the station server's own Icecast/SHOUTcast status endpoint off-thread. Cohesive with the existing radio-title methods already in this mixin (_radio_fetch_track_title / _radio_apply_track_title / radio_whats_playing); the endpoint parsing/derivation is the new under-cap quill/core/radio/station_status.py.",
-    "quill/ui/radio/player_controller.py": 605,
+    "quill/ui/radio/player_controller.py": 612,
+    "_rebaseline_2026_07_18_radio_engine_selection_logging": "player_controller.py 605 -> 612 (+7): quill-radio #5 observability, a debug line in _select_engine reporting the resolved backend (preference + mpv-present -> mpv/wx.media) so a 'why is this station on the wrong engine?' question has an answer in quill.log. Debug-level only; no behavior change.",
     "quill/ui/radio/station_browser_dialog.py": 696,
     "_rebaseline_2026_07_17_browse_stations_dropdowns": "station_browser_dialog.py 610 -> 696 (+86): quill-radio #2, country/genre become dropdowns instead of typo-prone free text. Tag/genre is now an editable wx.ComboBox and Country a wx.Choice with an (Any country) default, both filled off-thread from radio_browser.list_tags()/list_countries() and cached for the session (_populate_directory_choices / _apply_directory_choices), firing the search on selection. The pickable-vs-free-text choice is inherently this dialog's search UI; the pure country_query() helper is module-level and unit-tested.",
     "_rebaseline_2026_07_17_station_browser_iheart_tunein": "station_browser_dialog.py enters the tracked list at 610 (was under the 600 default): the iHeart + TuneIn source integration. The search fan-out (_do_search) now also blends TuneIn and iHeart results, a per-session iHeart sitemap index cache (_iheart_index), and each result row names its non-RadioBrowser source. The wx-free resolution/conversion helpers (tunein_search_stations / iheart_search_stations) live in the new under-cap quill/core/radio/directory_search.py, keeping this dialog's growth to just the UI wiring.",
